CRUD operations frontend added

+++ src/main/java/com/example/demo/controller/CourseController.java
@@ -6,6 +6,8 @@ import com.example.demo.service.CourseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import com.example.demo.service.InstructorService;
+
 
 import java.util.List;
 
@@ -17,10 +19,18 @@ public class CourseController {
     @Autowired
     private CourseService courseService;
 
+    @Autowired
+    private InstructorService instructorService;
+
   
     @GetMapping("/getCourses")
     public List<Course> getCourses() {
         return courseService.getAllCourses();
+    }
+
+    @GetMapping("/getCourse/{id}")
+    public Course getCourse(@PathVariable Long id) {
+        return courseService.getCourseById(id);
     }
 
     @PostMapping("/createCourse")
@@ -30,6 +40,12 @@ public class CourseController {
 
     @DeleteMapping("/deleteCourse/{id}")
     public void deleteCourse(@PathVariable Long id) {
+        Course course = courseService.getCourseById(id);
+        if (course == null) {
+            throw new RuntimeException("Course not found with id " + id);
+        }
+        Long instructor_id = course.getInstructorId();
+        instructorService.removeCourseFromInstructor(instructor_id, id);
         courseService.deleteCourse(id);
     }
 
@@ -50,8 +66,9 @@ public class CourseController {
 
     //delete file from a certain course
     @DeleteMapping("/{courseId}/deleteFile/{fileId}")
-    public void deleteCourseFile(@PathVariable Long courseId, @PathVariable Long fileId) {
+    public void deleteCourseFile(@PathVariable Long courseId ,@PathVariable Long fileId) {
         courseService.deleteCourseFile(courseId, fileId);
     }
 
+ 
 }
